requirements.txt
# OCR + PDF processing
pytesseract
pdf2image
Pillow
opencv-python
reportlab

# LLM + HTTP
requests
tenacity

# Observability
structlog
prometheus-client

# Environment variables
python-dotenv

# Utilities
numpy
fpdf
